+++ test/vanilla/Expected/AcceptanceTests/BodyInteger/bodyinteger/aio/_auto_rest_integer_test_service_async.py
@@ -22,7 +22,6 @@ class AutoRestIntegerTestService(object):
     :ivar int: IntOperations operations
     :vartype int: bodyinteger.aio.operations_async.IntOperations
     :param str base_url: Service URL
-    :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
     def __init__(