+++ test/vanilla/Expected/AcceptanceTests/BodyTime/bodytime/aio/_auto_rest_time_test_service_async.py
@@ -22,7 +22,6 @@ class AutoRestTimeTestService(object):
     :ivar time: TimeOperations operations
     :vartype time: bodytime.aio.operations_async.TimeOperations
     :param str base_url: Service URL
-    :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
     def __init__(