+++ test/vanilla/Expected/AcceptanceTests/Constants/constants/aio/_auto_rest_swagger_constant_service_async.py
@@ -22,7 +22,6 @@ class AutoRestSwaggerConstantService(object):
     :ivar contants: ContantsOperations operations
     :vartype contants: constants.aio.operations_async.ContantsOperations
     :param str base_url: Service URL
-    :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
     def __init__(