+++ test/vanilla/Expected/AcceptanceTests/MultipleInheritance/multipleinheritance/aio/_multiple_inheritance_service_client_async.py
@@ -20,7 +20,6 @@ class MultipleInheritanceServiceClient(MultipleInheritanceServiceClientOperation
     """Service client for multiinheritance client testing.
 
     :param str base_url: Service URL
-    :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
     """
 
     def __init__(